Implement timetable sharing as query string
